Check if generated files have to be regenerated in `refresh_generations.sh` to speed-up hook

+++ lib/generators/modulorails/githooks/templates/refresh_generations.sh
@@ -1,21 +1,29 @@
 #!/bin/sh
 
-echo 'Regenerate Gemfile.lock'
-bundle install
-git add Gemfile.lock
+modified_files=$(git diff --cached --name-only)
 
-if [ $(cat Gemfile.lock | grep i18n-js | wc -l) -gt 0 ]
+if echo "$modified_files" | grep -q "Gemfile.lock"
+then
+  echo 'Regenerate Gemfile.lock'
+  bundle install
+  git add Gemfile.lock
+fi
+
+if [ "$(cat Gemfile.lock | grep i18n-js | wc -l)" -gt 0 ] && echo "$modified_files" | grep -q "translations.js"
 then
   echo 'Regenerate JS translations'
   rake i18n:js:export
   git add app/assets/javascripts/i18n/translations.js
 fi
 
-echo 'Regenerate DB schema'
-export RAILS_ENV=test
-bundle exec rake db:drop db:create db:schema:load db:migrate
-git add db/schema.rb
-export RAILS_ENV=development
+if echo "$modified_files" | grep -q "schema.rb"
+then
+  echo 'Regenerate DB schema'
+  export RAILS_ENV=test
+  bundle exec rake db:drop db:create db:schema:load db:migrate
+  git add db/schema.rb
+  export RAILS_ENV=development
+fi
 
 if [ "$(git diff --cached --name-only | wc -l)" -ne 0 ]; then
   echo "Commit regenerated files by $GIT_AUTHOR_NAME <$GIT_AUTHOR_EMAIL>"
